Fix build

+++ src/video_core/renderer_vulkan/vk_rasterizer.cpp
@@ -81,7 +81,7 @@ VkViewport GetViewportState(const Device& device, const Maxwell& regs, size_t in
         height = -height;
     }
     if (lower_left) {
-        y = regs.surface_clip.height - (y + height);
+        y = static_cast<float>(regs.surface_clip.height) - (y + height);
         height = -height;
     }
     const float reduce_z = regs.depth_mode == Maxwell::DepthMode::MinusOneToOne ? 1.0f : 0.0f;
